+++ lib/build.sh
@@ -176,3 +176,10 @@ prune_devdependencies() {
     log_info "Successfully pruned devdependencies!"
   fi
 }
+
+warn_prebuilt_modules() {
+  local build_dir=$1
+  if [ -e "$build_dir/node_modules" ]; then
+    log_info "node_modules checked into source control" "https://devcenter.heroku.com/articles/node-best-practices#only-git-the-important-bits"
+  fi
+}